requirements.txt
# Data Science & Modeling
pandas
numpy
matplotlib
seaborn
statsmodels
scikit-learn
scipy
joblib
ipykernel

# Cloud & API Framework
fastapi
uvicorn
python-dotenv
